Update ThemeSwitch and change icon pack
