+++ mywatchlist-frontend/package.json
@@ -4,9 +4,7 @@
   "private": true,
   "homepage": ".",
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "@fortawesome/react-fontawesome": "^0.1.18",
+    "@heroicons/react": "^1.0.6",
     "@mui/icons-material": "^5.5.1",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
