+++ mywatchlist-frontend/src/components/Navbar/Navbar.tsx
@@ -1,6 +1,6 @@
 import { Component } from "react";
 import { Link } from "react-router-dom";
-import ToggleTheme from "../ToggleTheme/ToggleTheme";
+import ThemeSwitch from "../ThemeSwitch/ThemeSwitch";
 
 class Navbar extends Component {
     render() {
@@ -20,7 +20,7 @@ class Navbar extends Component {
                         </div>
 
                         <div className="flex items-center space-x-1">
-                            <ToggleTheme />
+                            <ThemeSwitch />
                             <Link to="/login" className="py-1 px-3 dark:text-white text-dark">Login</Link>
                             <Link to="/register" className="p-2 dark:text-white text-dark border border-emerald-300 rounded shadow">Signup</Link>
                         </div>
