+++ mywatchlist-frontend/tailwind.config.js
@@ -6,7 +6,8 @@ module.exports = {
         colors: {
           dark_navbar: "#0a192f",
           white_navbar: "#f0f0f0",
-          dark_bg: "#243247"
+          dark_bg: "#243247",
+          color_primary: "#64ffda"
         },
       }
     },
